Does not load story data upon creation
